refactoring and restrucuring of data
